fix: rename folder and move generator into it
