+++ cs/TagsCloudVisualization/Helpers/SizesGenerator.cs
